feat: remove duplications
