+++ PrebidMobileTests/RenderingTests/Tests/UtilitiesExtensionsTests/PBMConstantsTests.swift
@@ -66,12 +66,8 @@ class PBMConstantsTests: XCTestCase {
         XCTAssertEqual(GeoLocationConstants.DISTANCE_FILTER, 50.0)
     }
     
-    func testCloseButtonAreaConstant() {
-        XCTAssertEqual(PBMConstants.CLOSE_BUTTON_AREA_DEFAULT, 0.1)
-    }
-    
-    func testSkipButtonAreaConstant() {
-        XCTAssertEqual(PBMConstants.SKIP_BUTTON_AREA_DEFAULT, 0.1)
+    func testButtonAreaConstant() {
+        XCTAssertEqual(PBMConstants.BUTTON_AREA_DEFAULT, 0.1)
     }
     
     func testSkipDelayConstant() {
